Commonize v2 shim [databricks] (#4235)
* Commonize v2 shim


* Reuse some codes


* Revert "Reuse some codes"

This reverts commit 25549273d0abd7e1bd6a980e520224b597430a53.

* Reuse codes


* Revert a file
